+++ api/trading_test.go
@@ -272,6 +272,7 @@ func getTestGRPCServer(
 
 	sqlStore := sqlstore.SQLStore{}
 	sqlBalanceStore := sqlstore.NewBalances(&sqlStore)
+	sqlOrderStore := sqlstore.NewOrders(&sqlStore)
 
 	g := api.NewGRPCServer(
 		logger,
@@ -304,6 +305,7 @@ func getTestGRPCServer(
 		stakingService,
 		checkpointSvc,
 		sqlBalanceStore,
+		sqlOrderStore,
 	)
 	if g == nil {
 		err = fmt.Errorf("failed to create gRPC server")